Fixed issues: No Client Menu and Header in the Edit Page; Active client is skipped
